Moved Signals in a separate file.
